% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_selection.R
\name{clean_selection}
\alias{clean_selection}
\title{Clean selection}
\usage{
clean_selection(x)
}
\arguments{
\item{x}{Selected line(s) as character string}
}
\description{
Clean selection
}
\keyword{internal}
