% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke_viewer.R
\name{invoke_viewer}
\alias{invoke_viewer}
\title{Invoke the RStudio Data Viewer on Highlighted Text}
\usage{
invoke_viewer()
}
\description{
Invoke the RStudio Data Viewer on Highlighted Text
}
\details{
This function will invoke the RStudio Data Viewer on highlighted text. Unlike built-in functions like F2 that invoke the viewer on a highlighted object name, this function will additionally accept piped sequences of commands. After installing the package, go to "Modify Keyboard Shortcuts" and set a shortcut of your choice for the name "Invoke Viewer".
}
