% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_selection.R
\name{parse_selection}
\alias{parse_selection}
\title{Parse selection}
\usage{
parse_selection(x)
}
\arguments{
\item{x}{Cleaned line(s) as character string}
}
\description{
Parse selection
}
\keyword{internal}
