% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_selection.R
\name{send_selection}
\alias{send_selection}
\title{Send selection to console}
\usage{
send_selection(selection_cleaned, selection_parsed)
}
\arguments{
\item{selection_cleaned}{Cleaned line(s) as character string}

\item{selection_parsed}{A `data.frame` as returned by `parse_selection()`}
}
\description{
Send selection to console
}
\keyword{internal}
