% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_selection.R
\name{view_selection}
\alias{view_selection}
\title{View selected lines in RStudio data viewer}
\usage{
view_selection()
}
\description{
This function unfolds its full potential when it's mapped to a
  sensible keyboard shortcut: In RStudio, navigate to 'Tools' > 'Modify
  Keyboard Shortcuts' and then search for 'View current selection'. Click on
  the 'Shortcut' field to assign a new keyboard shortcut.
}
\seealso{
For a more detailed overview and a demo, please check the
  \href{https://github.com/lukasfeick-sw/viewR}{GitHub repository}.
}
