% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewr_widget.R
\name{viewdt-shiny}
\alias{viewdt-shiny}
\alias{viewdtOutput}
\alias{renderViewdt}
\title{Shiny bindings for viewdt}
\usage{
viewdtOutput(outputId, width = "100\%", height = "600px")

renderViewdt(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Must be valid CSS units (e.g. `"100\%"`) or numbers.}

\item{expr}{An expression that generates a [viewdt()] widget.}

\item{env}{The environment in which to evaluate `expr`.}

\item{quoted}{Is `expr` a quoted expression (with `quote()`)?}
}
\value{
`viewdtOutput()` returns a Shiny output UI element;
  `renderViewdt()` returns a Shiny render function.
}
\description{
Output and render functions for using [viewdt()] within Shiny applications
and interactive R Markdown documents.
}
